Optimize and Python 3 cleanup
- Fix shutil.rmtree('tmp') crash in setup.py
- Remove eval(input()) security vulnerability
- Replace base64.decodestring() with decodebytes()
- Replace .encode('hex') with .hex()
- Fix wpa_cli bytes/str comparison in wpspbc.py
- Remove all from __future__ imports
- Fix setup.py classifiers: Python 2 -> Python 3 :: Only
- Replace import mock with from unittest import mock
- Remove all # pylint: skip-file comments
- Replace open(os.devnull) with subprocess.DEVNULL
- Fix resource leaks (victim.py url_file, sniffer.py DN, handshakeverify.py keyfile)
- Migrate metadata to PEP 621 pyproject.toml with uv sources
- Add pyric to declared dependencies
- Replace deprecated distutils imports with setuptools._distutils
- Shrink setup.py to build-only checks
- Replace install.sh with uv-based workflow
- Add Makefile for dev workflow
- Replace .travis.yml with GitHub Actions CI using uv
- Re-enable test_interfaces.py with conditional skip
- Update README with uv-based install instructions

+++ setup.py
@@ -1,46 +1,16 @@
 #!/usr/bin/env python3
-r"""
-                    _  __ _       _     _     _
-                   (_)/ _(_)     | |   (_)   | |
- ((.))    __      ___| |_ _ _ __ | |__  _ ___| |__   ___ _ __
-   |      \ \ /\ / / |  _| | '_ \| '_ \| / __| '_ \ / _ \ '__|
-  /_\      \ V  V /| | | | | |_) | | | | \__ \ | | |  __/ |
- /___\      \_/\_/ |_|_| |_| .__/|_| |_|_|___/_| |_|\___|_|
-/     \                    | |
-                           |_|  Version {}
+"""
+Build-time checks for wifiphisher dependencies.
 """
 
 import os
 import sys
 import shutil
 import tempfile
-import distutils.sysconfig
-import distutils.ccompiler
-
-from distutils.errors import CompileError, LinkError
-from setuptools import Command, find_packages, setup
 from textwrap import dedent
-
-try:
-    raw_input  # Python 2
-    sys.exit("Please use Python 3 to install Wifiphisher.")
-except NameError:
-    pass  # Python 3
-
-
-class CleanCommand(Command):
-    """Custom clean command to tidy up the project root."""
-
-    user_options = []
-
-    def initialize_options(self):
-        pass
-
-    def finalize_options(self):
-        pass
-
-    def run(self):
-        os.system("rm -vrf ./build ./dist ./*.pyc ./*.tgz ./*.egg-info")
+from setuptools import setup
+from setuptools._distutils import sysconfig, ccompiler
+from setuptools.errors import CompileError, LinkError
 
 
 # code for checking if libnl-dev and libnl-genl-dev exist
@@ -71,22 +41,14 @@ LIBNAME_CODE_DICT = {"netlink": LIBNL_CODE, "openssl": OPENSSL_CODE}
 
 
 def check_required_library(libname, libraries=None, include_dir=None):
-    """
-    Check if the required shared library exists
-
-    :param libname: The name of shared library
-    :type libname: str
-    :return True if the required shared lib exists else false
-    :rtype: bool
-    """
     build_success = True
     tmp_dir = tempfile.mkdtemp(prefix="tmp_" + libname + "_")
     bin_file_name = os.path.join(tmp_dir, "test_" + libname)
     file_name = bin_file_name + ".c"
     with open(file_name, "w") as filep:
         filep.write(LIBNAME_CODE_DICT[libname])
-    compiler = distutils.ccompiler.new_compiler()
-    distutils.sysconfig.customize_compiler(compiler)
+    compiler = ccompiler.new_compiler()
+    sysconfig.customize_compiler(compiler)
     try:
         compiler.link_executable(
             compiler.compile([file_name], include_dirs=include_dir),
@@ -118,13 +80,6 @@ def check_required_library(libname, libraries=None, include_dir=None):
 
 
 def check_dnsmasq():
-    """
-    Try to install dnsmasq on host machine if not present.
-
-    :return: None
-    :rtype: None
-    """
-
     if not os.path.isfile("/usr/sbin/dnsmasq"):
         sys.exit(
             "dnsmasq not found in /usr/sbin/dnsmasq. "
@@ -134,64 +89,14 @@ def check_dnsmasq():
         )
 
 
-# setup settings
-NAME = "wifiphisher"
-AUTHOR = "sophron"
-AUTHOR_EMAIL = "sophron@latthi.com"
-URL = "https://github.com/wifiphisher/wifiphisher"
-DESCRIPTION = "Automated phishing attacks against Wi-Fi networks"
-LICENSE = "GPL"
-KEYWORDS = ["wifiphisher", "evil", "twin", "phishing"]
-PACKAGES = find_packages(exclude=["docs", "tests"])
-INCLUDE_PACKAGE_DATA = True
-VERSION = "1.4"
-CLASSIFIERS = [
-    "Development Status :: 5 - Production/Stable",
-    "License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)",
-    "Natural Language :: English",
-    "Operating System :: Unix",
-    "Programming Language :: Python :: 2",
-    "Programming Language :: Python :: 2.7",
-    "Programming Language :: Python :: 2 :: Only",
-    "Topic :: Security",
-    "Topic :: System :: Networking",
-    "Intended Audience :: End Users/Desktop",
-    "Intended Audience :: System Administrators",
-    "Intended Audience :: Information Technology",
-]
-ENTRY_POINTS = {"console_scripts": ["wifiphisher = wifiphisher.pywifiphisher:run"]}
-INSTALL_REQUIRES = ["pbkdf2", "scapy", "tornado>=5.0.0", "pyric"]
-DEPENDENCY_LINKS = [
-    "http://github.com/sophron/pyric/tarball/master#egg=pyric-0.5.0",
-]
-CMDCLASS = {
-    "clean": CleanCommand,
-}
 LIB_NL3_PATH = "/usr/include/libnl3"
 LIB_SSL_PATH = "/usr/include/openssl"
 
 check_dnsmasq()
 check_required_library("netlink", ["nl-3", "nl-genl-3"], [LIB_NL3_PATH])
 check_required_library("openssl", ["ssl"], [LIB_SSL_PATH])
-shutil.rmtree("tmp")
 
-# run setup
-setup(
-    name=NAME,
-    author=AUTHOR,
-    author_email=AUTHOR_EMAIL,
-    description=DESCRIPTION,
-    license=LICENSE,
-    keywords=KEYWORDS,
-    packages=PACKAGES,
-    include_package_data=INCLUDE_PACKAGE_DATA,
-    version=VERSION,
-    entry_points=ENTRY_POINTS,
-    install_requires=INSTALL_REQUIRES,
-    dependency_links=DEPENDENCY_LINKS,
-    classifiers=CLASSIFIERS,
-    url=URL,
-    cmdclass=CMDCLASS,
-)
+if os.path.exists("tmp"):
+    shutil.rmtree("tmp")
 
-print(__doc__.format(VERSION))  # print the docstring located at the top of this file
+setup()
